+++ test/dummy/config/initializers/recording_studio_oauth.rb
@@ -10,3 +10,7 @@ RecordingStudioOauth.configure do |config|
   config.register_origin_as_protected_resource = false
   config.extra_protected_resource_paths = []
 end
+
+RecordingStudioShopifyPluginTemplate.configure do |config|
+  config.registered_app_client_id = ENV["SHOPIFY_PLUGIN_REGISTERED_APP_CLIENT_ID"]
+end